Generate unique cert serials from a db sequence.
Changed the BigInteger to Long to appease Hibernate, if we need more
certs than that we've got storage issues anyhow.

+++ proxy/src/test/java/org/fedoraproject/candlepin/service/impl/stub/StubEntitlementCertServiceAdapter.java
@@ -15,7 +15,6 @@
 package org.fedoraproject.candlepin.service.impl.stub;
 
 import java.io.IOException;
-import java.math.BigInteger;
 import java.security.GeneralSecurityException;
 import java.util.Date;
 
@@ -50,7 +49,7 @@ public class StubEntitlementCertServiceAdapter extends BaseEntitlementCertServic
     @Override
     public EntitlementCertificate generateEntitlementCert(Consumer consumer,
         Entitlement entitlement, Subscription sub, Product product, Date endDate, 
-        BigInteger serialNumber) throws GeneralSecurityException, IOException {
+        Long serialNumber) throws GeneralSecurityException, IOException {
         log.debug("Generating entitlement cert for:");
         log.debug("   consumer: " + consumer.getUuid());
         log.debug("   product: " + product.getId());
